one script build

+++ server/Controllers/HomeController.js
@@ -6,13 +6,15 @@ import BaseController from './BaseController';
 import Home from '../../src/pages/Home';
 import {InfoController, EventController} from './index';
 import models from '../../models';
+import bcrypt from "bcrypt-nodejs";
+import passport from "passport/lib";
 
 export default class HomeController extends BaseController {
 
     static routing() {
         return {
             cassiopeia_home: {type:'get', route:'/',action: 'indexAction'},
-            // danger: {type:'get', route:'/danger',action: 'danger'}
+            danger: {type:'get', route:'/danger',action: 'danger'}
         };
     }
 
@@ -40,12 +42,23 @@ export default class HomeController extends BaseController {
     }
 
 
-    danger() {
-        models.User.sync({force:true}).then(() => {
-            models.Event.sync({force: true});
-            models.Info.sync({force:true});
-            models.Contact.sync({force:true});
-        });
+    danger(req) {
+        if(!!+req.query.force) {
+            models.User.sync({force:true}).then(() => {
+                models.Event.sync({force: true});
+                models.Info.sync({force:true});
+                models.Contact.sync({force:true});
+                models.Photo.sync({force:true});
+
+                models.User.create({
+                    username: 'admin',
+                    firstName: 'admin',
+                    lastName: 'admin',
+                    password: bcrypt.hashSync('admin'),
+                    roles: ['admin']
+                })
+            });
+        }
     }
 }
